requirements.txt
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
sentry-sdk==2.20.0
